requirements/base.txt
Django==3.0.6
django-environ==0.4.5
psycopg2-binary==2.8.5
Pillow==7.1.2
djangorestframework==3.11.0
environ==1.0

Celery==4.4.5
django-redis==4.12.1   # https://github.com/jazzband/django-redis
djangorestframework-jwt==1.11.0



